iwlwifi: mvm: support HE TXing with MCS10 and MCS11
Allow setting an HE rate of MCS10 and MCS11 for TXing.

Note that this patch does not affect the rate scaling
whatsoever, and the driver will still not reach HE
rates without setting it as a fixed rate. However, as
this is expected to be done by the FW, it is good
enough for now.

type=feature
ticket=jira:WLWG-3845

Change-Id: Ic82c684b799e43a1c9f56f48738666a265e77983
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/126246
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: bbb17628b2e86dae6d788c1620f6cc977a616bbc