[BUGFIX]mac80211_hwsim: Use proper TX power
In struct ieee80211_tx_info, control.vif pointer and rate_driver_data[0]
falls on the same place, depending on the union usage.
During the whole TX process, the union is referred to as a control struct,
which holds the vif that is later used in the tx flow, especially in order
to derive the used tx power.
Referring direcly to rate_driver_data[0] and assigning a value to it,
overwrites the vif pointer, hence making all later references irrelevant.
Moreover, rate_driver_data[0] isn't used later in the flow in order to
retrieve the channel that it is pointing to.

type=bugfix
ticket=none
fixes=unknown

Change-Id: I8cce5d6465a5f7ae5cdc19ea55f0a9b32a51bd0b
Signed-off-by: Beni Lev <beni.lev@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/128508
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 04e11835a7d42fa82766f35668287ad19bbea4e9