[BUGFIX] iwlwifi: mvm: remove session protection to allow channel switch
If a time event is already scheduled when trying to schedule one for
channel switch, the code assumes the channel switch is already
scheduled and no further action is required.
However, it is possible that the scheduled time event is actually
for session protection (e.g. when the first beacon after association
contains the CSA IE). In this case the channel switch will not be
scheduled which will finally lead to disconnection.

Fix this by removing the old time event and schduling a new one for
the channel switch.

type=bugfix
ticket=jira:WIFILNX-763
fixes=unknown

Change-Id: I6d2a3251247e28d564411e6d1f057c394be73426
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/130615
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Cohen, Guy <guy.cohen@intel.com>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: 8cc4a2b2b9a3cc7965eef6eca0266d85e0a72724