iwlwifi: acpi: add common code to read from ACPI
There are many places where the same process of invoking a method from
ACPI is used, causing a lot of duplicate code.  To improve this,
introduce a new function to get an ACPI object by invoking an ACPI
method that can be reused.

Additionally, since this function needs to be called when we only have
the trans, the opmode or the device, introduce a new debug macro that
gets the device as a parameter so it can be used in the new function.

type=cleanup
ticket=jira:WLWG-4460

Change-Id: I3867d02a1ab753d59a21e4e2f882b37cfe2a41a3
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/139299
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Dreyfuss, Haim <haim.dreyfuss@intel.com>
x-iwlwifi-stack-dev: ea7fc707a5b60d912b27a63a58852e796da2c4c2