scripts/kernel-doc: warn on excess enum value descriptions
The existing message
	"Excess struct/union/enum/typedef member [...]"
made it sound like this would already be done, but the
code is never invoked for enums or typedefs (and really
can't be).

Add some code to the enum dumper to handle this there
instead.

While at it, also make the above message more accurate
by simply dumping the type that was passed in, and pass
the struct/union differentiation in.

type=cleanup
ticket=none

Change-Id: I5ce316087accb9fe18d3032b4eaddc5cfa77f368
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/138847
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: a16fa0d8953495a69c4ccdae1eadb5cd57356471