iwlwifi: xvt: add common driver command
Add common driver command. This command is meant to serve as a wrapper
for commands with minimal implementation in the USC. It allows
the USC to serve merely as a pipe for tool<->driver/FW tests. It is
added as part of 'pyfi' tests requirements.

type=feature
ticket=none

Change-Id: I4e675b11a93ae51d09bb05204c985260e042dedc
Signed-off-by: Nitzan Kugelman <nitzanx.kugelman@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/140883
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: 83d1c69ad0a647818d4413c91551f092da3ae054