[BUGFIX] mac80211: Call mgd_prep_tx before transmitting deauthentication
In multi channel scenarios, when disassociating from the AP before a
beacon was heard from the AP, it is not guaranteed that the virtual
interface is granted air time for the transmission of the
deauthentication frame. This in turn can lead to various issues as
the AP might never get the deauthentication frame.

To mitigate such possible issues, add a HW flag indicating that the
driver requires mac80211 to call the mgd_prep_tx() driver callback
to make sure that the virtual interface is granted immediate airtime
to be able to transmit the frame, in case that no beacon was heard
from the AP.

type=bugfix
ticket=jira:WIFI-7273
fixes=unknown

Change-Id: I6d0e89d7dd1fe218ba385fb99ceee737f0bcb085
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/157336
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: 337f4b065ba9bdf1870776d405800e5dbbd517f3