iwlwifi: mvm: warn if seq_ctrl gets out of sync
In 22560 HW, the FW is the one to actually set the
sequence number. However, it is critical for the
driver to know the current sequence number and for
it to not get out of sync with the FW. (For example,
this is required when the driver specifies the seq
to start a BA session on.)

To help us catch bugs early, add a warning if the
seq num the driver used is different from the one
the FW set.

type=feature
ticket=none

Change-Id: Ia5b44e0abc042f9dfebb6ff5ae96f4f6a5e5d4ce
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/159982
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: 84e8e12050669740939b9d6725bcba1ffe2a525c