iwlwifi: pcie: allow sending pre-built A-MSDUs
In case of A-MSDUs, the trans layer is taking care of building
the subframes (out of the given skb), according to the given gso_size.

However, in some testing flows, we want to build the whole A-MSDU
frame in a different place (e.g. userspace), and ask the driver
to send it as-is.

In case of gso_size==0, simply treat the frame as normal-frame,
although the A-MSDU flag is set.

type=feature
ticket=none

Change-Id: I1090bfeeecddd275da437277bdc6f0970c8a9a5c
Signed-off-by: Eliad Peller <eliadx.peller@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/160657
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Sharon, Sara <sara.sharon@intel.com>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: b13845ceb8a705ff4c210721720fe0807d02fe50