[BUGFIX] iwlwifi: add shared clock PHY config flag for some devices
Some devices use a shared clock which is very sensitive to variations
and cause trouble in some situations.  We need to set a bit in the phy
configuration to indicate that to the FW.  To make this generic, add a
extra_phy_config_flags element to the device configuration and OR it
into the phy_cfg before sending it to the firmware.  And also create a
set of configurations for devices that use shared clocks and need this
extra bit to be set.

type=bugfix
ticket=jira:WIFI-5267
fixes=Ia556e57174640ab2b6e0039ae15afa9dbdf6c08a

Change-Id: Ia79d6633dd7a9cb483ef41c4eee4cf1cf307ce12
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: d88098e12faa750fc6e57e7e99eb29d4ca69f3cf