[BUGFIX] iwlwifi: mvm: don't assign tx queue id for a reserved queue
When starting tx aggregation for a tid that does not yet have an
allocated queue, a new queue is reserved for the tid. The reserved
queue id is immediately assigned to the tid_data, although the queue
is not yet allocated.
As a result, all following tx on this tid will fail because the queue
is not allocated.

Fix this by not updating the tid_data yet, which will result in
actually allocating the reserved queue when trying to tx on this tid.

type=bugfix
ticket=jira:WIFI-8161
fixes=unknown

Change-Id: Id4cafc1e717692f6d1cb28a6a57338ced762a274
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/164654
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: e6e853da73b2f9db523e7a6913a4d7e27759658a