[BUGFIX] iwlwifi: set the tid for non-QOS frames to zero
The tid for Non-QOS frames is set to IWL_MAX_TID_COUNT. This value
is also used for configuring the queue for non-QOS data. However, this
tid is used by the FW for management queues. As a result, the FW
does not encrypt non-QOS data frames.

Fix this by setting the tid for non-QOS data frames to zero, which
is a valid value for data frames in the FW.

This also fixes a bug in sending multicast frames, where the queues
are allocated with tid == 0, but are sent with tid == 8, which may
lead to unexpected behavior.

type=bugfix
ticket=none
fixes=unknown

Change-Id: Ib0eb0a80079b0aa8183a084e5bce55501e0abf23
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
x-iwlwifi-stack-dev: 75f3b8eb70c2a76161fad22bebe8efc28b050b7e