iwlwifi: remove TCM Kconfig flag
Originally we added all TCM code behind a Kconfig flag because
implementing this in mac80211 was being considered for upstream.  But
it turned out that implementing it in mac80211 would be too
complicated, so it never happened.  Now we decided that we can
upstream the entire TCM code in the driver, so there's no need to keep
it flagged with the IWLMVM_TCM Kconfig flag.

Remove the flag and unifdef all code with the flag set.

type=cleanup
ticket=jira:WLWG-4818

Change-Id: Ief61b02a7136dab21b3ba06c49378cd897fbb266
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: a5326a98363c6726bc7d00bfe4c223c7b964a999