[BUGFIX] iwlwifi: pcie: set interrupt coalescing also for gen2
We offloaded all the RX configuration of init to firmware. However,
the configuration of interrupt coalescing was left hanging - it wasn't
offloaded nor was it written by host.

This write to the CSR is allowed in gen2, so the host can do it.
Without it we have various issues with RX fullness.

type=bugfix
fixes=unknown
ticket=none

Change-Id: I5a3668228deb29361b06fb485443fb6d335f8c52
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
x-iwlwifi-stack-dev: 3025a7a8369c9718e0c11198d1ffe7f6763e23b4