[BUGFIX] iwlwifi: fix devices with PCI Device ID 0x34F0 and 11ac RF modules
The devices with PCI device ID 0x34F0 are part of the SoC and can be
combined with some different external RF modules.  The configuration
for these devices should reflect that, but are currently mixed up.  To
avoid confusion with discrete devices, add part of the firmware to be
used and the official name of the device to the cfg structs.

This is least reorganization possible (without messing things even
more) that could be done as a bugfix for this SoC.  Further
reorganization of this code will be done separately.

type=bugfix
fixes=unknown
ticket=none

Change-Id: I2d442817590366dcf3f02ac080017c44fa4d0e95
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>x-iwlwifi-stack-dev: 3be00e42fe68c49d3daab9ead6df0cbd512f4cb3