iwlwifi: wrt: add to dump number of lmacs, lmac1 and umac error id
Add to the dump the number of lmacs, the error id of the umac
and the error id of lmac1, if supported.
In case the reason for the dump trigger is not an assert
the error id is zero.

type=feature
ticket=none

Change-Id: I69cdd6423103eba607965c6b6d7e7e8614fa1cff
Signed-off-by: Shahar S Matityahu <shahar.s.matityahu@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/190009
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 4596fbbf9ca4150ffc7702e07f4ca1f8450ea827