[BUGFIX] iwlwifi: don't replace 22260 config with 22000
In the trans.c code, we check for the RF_ID of the device and hardcode
the config to iwl22000_2ax_cfg_qnj*, depending on the radio version.
This should only be done for 22000 devices, but it is currently
overwriting 22260's configuration as well.  Fix it by skipping this
code if the configuration is for 22260.

type=bugfix
fixes=Ibcf4f5d896a738fde0c3e2b0d3db62763c970f6c
ticket=none

Change-Id: Ia6eca8ccba3e2cb276f705ab0cf31c3bbbeadb6d
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/191933
Tested-by: Givon, Oren <oren.givon@intel.com>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 98dd42a4656bcfadb73a06f66dd929dc14be29d4