[BUGFIX] iwlwifi: pcie: fix emergency path
Allocator swaps the pending requests with 0 when it starts
working. This means that relying on it n RX path to decide if
to move to emergency is not always a good idea, since it may
be zero, but there are still a lot of unallocated RBs in the
system. Change allocator to decrement the pending requests on
real time. It is more expensive since it accesses the atomic
variable more times, but it gives the RX path a better idea
of the system's status.

type=bugfix
fixes=I711058183ae488863cab6d6df172142d2c805594
ticket=none

Change-Id: I7e2a509d9a2765d332e4d18e7c9ea1e438fe47ed
Reported-by: Ilan Peer <ilan.peer@intel.com>
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/202190
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 2dadcc5dd8d9728b92512dba81cbd12a43841654