mac80211_hwsim: make copying of ciphers safer by checking the length
Make sure the length of the ciphers we are copying never exceeds the
space we have for storing them.  There is no risk of overcopying at
the moment, because we check n_params before, but this makes this
function safer in case someone changes something in the future.

type=cleanup
ticket=none

Change-Id: Icaea2b45a78d2e939503b7c57c90670501274219
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>x-iwlwifi-stack-dev: dfb043da588b97dc7c62b3fe9e445db672ea3adb