backports: handle RHEL 7.6 kernel
RHEL 7.6 uses kernel v3.10 with a lot of backports from newer
kernels.  This causes compilation to break in many places when the
backports project is used.  To solve the issue, check for RHEL release
code 7.6 where appropriate.  This doesn't take into consideration
lower versions of the kernel (i.e. < 7.6), because it's very difficult
to find the information about when each feature was backported.

type=maint
ticket=none

Change-Id: Ic310f826e4d1e036eca65b4fa7d13d2da12620aa
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>x-iwlwifi-stack-dev: a20591e272e3a6aa3614fb091fb7a545958870f1