[NOUPSTREAM] backport: rename ktime_get_boot_ns for v5.3+
In v5.3, this function was renamed, so rename it here for v5.3+.  When
we merge v5.3 back from upstream, the opposite should be done (i.e. we
will have _boottime_ and need to rename to _boot_ in < v5.3 instead).

This may be applied to older cores if we want them to compile on
v5.3+.

type=maint
ticket=none

Change-Id: Ia3f2b514a173edb320b3445a9d6e7e31e9182059
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/233392
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: eb274c49a30f3bf2cf3ca6f537a018707edc0bae