iwlwifi: convert the 9260-1x1 device to use the new parameters
The 9260-1x1 device can be differentiated using the PCI device ID.
There is a single occurrence of this device, so continue relying on
the device and subsystem device IDs.

The name of this device was incorrect, so add a new string
specifically for it.

type=feature
ticket=jira:WIFI-18929

Change-Id: Id27bd9c878b73cb771691cbe6082fd40e079b44d
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/241487
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 7fe012073f94ed6b2cb685d32f28d13ef9480085