cfg80211: send scan req for 6/7 ghz band
After the first scan on the legacy bands is done,
iterate over the BSS entries found in the last scan and:
1. Parse the RNR IE
2. Iterate the collocated APs list and add each AP to the second scan
request. If this is not the first AP added for the channel, just
add the new bssid and short ssid to the exiting channel.
3. Anyway, add all the PSC channels (if it was not reported by the RNR
already).

type=feature
ticket=jira:WIFI-31020

Change-Id: I956f91022ef81544688a61adea963fb6912a192e
Signed-off-by: Tova Mussai <tova.mussai@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/237454
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 0d9006444aa953e75804deb145b59fb54818b2d9