[NOUPSTREAM] iwlwifi: yoyo: use firmware_request_nowarn() to load iwl-dbg-cfg.ini
If the iwl-dbg-cfg.ini file is not available, we always get a warning
from the FW loading subsystem that the file is not available, which is
okay, but can be confusing.  Replace the request_firmware() call with
firmware_request_nowarn() call to suppress it.  Note that this only
works with v4.18 or above.  On older kernels, the message will still
be there.

type=feature
ticket=jira:WIFI-38641

Change-Id: Ibbddb46571c5f3abbf3c856b008c6f001f352a7b
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/259507
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 5e798387f7926e5a3a934d018aa46fd07b00c452